requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore==2.16.0
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.2.1
httpx==0.27.2
